Remove ODataAnnotatableExtensions
